# Remove merge from Compare