--- a/internal/testdata/scalar-compare-output.txt
+++ b/internal/testdata/scalar-compare-output.txt
@@ -1,4 +1,2 @@
-- Merged Object:
-{types=[{name="scalar";scalar="numeric"}]}
 - Modified Fields:
 .types[name="scalar"].scalar