--- a/typed/typed.go
+++ b/typed/typed.go
@@ -18,6 +18,7 @@ package typed
 
 import (
 	"fmt"
+	"strings"
 	"sync"
 
 	"sigs.k8s.io/structured-merge-diff/fieldpath"
@@ -116,7 +117,7 @@ func (tv TypedValue) Compare(rhs *TypedValue) (c *Comparison, err error) {
 		Modified: fieldpath.NewSet(),
 		Added:    fieldpath.NewSet(),
 	}
-	c.Merged, err = merge(&tv, rhs, func(w *mergingWalker) {
+	_, err = merge(&tv, rhs, func(w *mergingWalker) {
 		if w.lhs == nil {
 			c.Added.Insert(w.path)
 		} else if w.rhs == nil {
@@ -126,8 +127,6 @@ func (tv TypedValue) Compare(rhs *TypedValue) (c *Comparison, err error) {
 			// Need to implement equality check on the value type.
 			c.Modified.Insert(w.path)
 		}
-
-		ruleKeepRHS(w)
 	}, func(w *mergingWalker) {
 		if w.lhs == nil {
 			c.Added.Insert(w.path)
@@ -268,10 +267,6 @@ func merge(lhs, rhs *TypedValue, rule, postRule mergeRule) (*TypedValue, error) 
 // No field will appear in more than one of the three fieldsets. If all of the
 // fieldsets are empty, then the objects must have been equal.
 type Comparison struct {
-	// Merged is the result of merging the two objects, as explained in the
-	// comments on TypedValue.Merge().
-	Merged *TypedValue
-
 	// Removed contains any fields removed by rhs (the right-hand-side
 	// object in the comparison).
 	Removed *fieldpath.Set
@@ -289,15 +284,15 @@ func (c *Comparison) IsSame() bool {
 
 // String returns a human readable version of the comparison.
 func (c *Comparison) String() string {
-	str := fmt.Sprintf("- Merged Object:\n%v\n", c.Merged.AsValue())
+	bld := strings.Builder{}
 	if !c.Modified.Empty() {
-		str += fmt.Sprintf("- Modified Fields:\n%v\n", c.Modified)
+		bld.WriteString(fmt.Sprintf("- Modified Fields:\n%v\n", c.Modified))
 	}
 	if !c.Added.Empty() {
-		str += fmt.Sprintf("- Added Fields:\n%v\n", c.Added)
+		bld.WriteString(fmt.Sprintf("- Added Fields:\n%v\n", c.Added))
 	}
 	if !c.Removed.Empty() {
-		str += fmt.Sprintf("- Removed Fields:\n%v\n", c.Removed)
+		bld.WriteString(fmt.Sprintf("- Removed Fields:\n%v\n", c.Removed))
 	}
-	return str
+	return bld.String()
 }