Speed is not super reliable on laptop because of throttling, allocs should be relevant:
```
benchmark                                             old ns/op     new ns/op     delta
BenchmarkDeducedSimple-8                              103885        86064         -17.15%
BenchmarkDeducedNested-8                              256101        247159        -3.49%
BenchmarkDeducedNestedAcrossVersion-8                 259720        286096        +10.16%
BenchmarkLeafConflictAcrossVersion-8                  85375         89451         +4.77%
BenchmarkMultipleApplierRecursiveRealConversion-8     733484        720364        -1.79%

benchmark                                             old allocs     new allocs     delta
BenchmarkDeducedSimple-8                              877            840            -4.22%
BenchmarkDeducedNested-8                              2465           2354           -4.50%
BenchmarkDeducedNestedAcrossVersion-8                 2579           2422           -6.09%
BenchmarkLeafConflictAcrossVersion-8                  866            812            -6.24%
BenchmarkMultipleApplierRecursiveRealConversion-8     5845           5680           -2.82%

benchmark                                             old bytes     new bytes     delta
BenchmarkDeducedSimple-8                              67670         63922         -5.54%
BenchmarkDeducedNested-8                              156494        145361        -7.11%
BenchmarkDeducedNestedAcrossVersion-8                 164866        148804        -9.74%
BenchmarkLeafConflictAcrossVersion-8                  69100         63307         -8.38%
BenchmarkMultipleApplierRecursiveRealConversion-8     570178        557684        -2.19%
```